requirements.txt
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.24.0
python-multipart==0.0.6
llmsherpa==0.1.3
cohere==4.34
numpy==1.26.1
beautifulsoup4==4.12.2
urllib3==2.0.7
lxml==4.9.3
supabase==2.0.3
python-dotenv==1.0.0
openai==1.3.5
